Refactor NguoiDung model: Remove unused fields
Remove the 'diaChi' and 'dienThoai' fields from the NguoiDung model as they are no longer needed.

Refactor HoaDonDTO and SanPhamChiTietDTO: Organize imports

Clean up the imports in the HoaDonDTO and SanPhamChiTietDTO classes to remove unused imports and organize them properly.

Refactor GioHangChiTietDTO: Change price type to BigDecimal

Change the 'giaTungSanPham' field in the GioHangChiTietDTO class from Double to BigDecimal to ensure accurate representation of prices.

Refactor KhuyenMaiRepository: Reorder imports

Reorder the imports in the KhuyenMaiRepository class to follow the established convention.

Refactor ChiTietSanPham model: Organize imports

Clean up the imports in the ChiTietSanPham class to remove unused imports and organize them properly.

Refactor DiaChiRepository: Reorder imports

Reorder the imports in the DiaChiRepository class to follow the established convention.

Refactor ChiTietGioHangRepository: Reorder imports

Reorder the imports in the ChiTietGioHangRepository class to follow the established convention.

Refactor KhuyenMai model: Organize imports

Clean up the imports in the KhuyenMai class to remove unused imports and organize them properly.

Refactor GioHangService: Organize imports

Clean up the imports in the GioHangService class to remove unused imports and organize them properly.

Refactor ChiTietGioHang model: Change price type to BigDecimal

Change the 'giaTungSanPham' field in the ChiTietGioHang class from Double to BigDecimal to ensure accurate representation of prices.

Refactor ChatLieuVoService: Organize imports

Clean up the imports in the ChatLieuVoService class to remove unused imports and organize them properly.

Refactor HoaDonController: Organize imports

Clean up the imports in the HoaDonController class to remove unused imports and organize them properly.

+++ src/main/java/com/example/demo/controller/GioHangController.java
@@ -1,6 +1,7 @@
 package com.example.demo.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.math.BigDecimal;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -8,18 +9,17 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.model.GioHang;
-import com.example.demo.repository.ChiTietGioHangRepository;
 import com.example.demo.repository.GioHangRepository;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.PostMapping;
 
 @RestController
 @RequestMapping("/api/carts")
@@ -55,7 +55,7 @@ public class GioHangController {
     @PostMapping
     public ResponseEntity<GioHang> saveCart(@RequestBody GioHang gioHang) {
         gioHang.setId(Long.valueOf(0));
-        gioHang.setTongGia(Double.valueOf(0));
+        gioHang.setTongGia(BigDecimal.valueOf(0));
         gioHang.setTongSoLuong(Integer.valueOf(0));
         return ResponseEntity.ok(gioHangRepository.save(gioHang));
     }
