+++ src/main/java/com/example/demo/model/ChiTietGioHang.java
@@ -1,10 +1,9 @@
 package com.example.demo.model;
-import com.fasterxml.jackson.annotation.JsonBackReference;
+import java.math.BigDecimal;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -12,7 +11,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -40,7 +38,7 @@ public class ChiTietGioHang {
     private Integer soLuong;
 
     @Column(name = "Gia_tung_san_pham")
-    private Double giaTungSanPham;
+    private BigDecimal giaTungSanPham;
 
     // Getters and Setters
 }
